Parser - Fix flattening, improve regex builder
Flattening problem - Lexer rules that are referenced from both parser and lexer rules were not present in the grammar. They were flattened inside a recursion as a part of another lexer rule but not added to resulting GrammarInfo rules.

Regex builder now escapes characters inside literals that will be used inside a regex. Unnecessary brackets are not inserted into the regex.